backend/pyproject.toml
[tool.poetry]
name = "voucher pdf generator"
version = "0.1.0"
description = ""
authors = ["Захаров Павел <pzakharov83@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
sqlalchemy = "^1.4.46"
jinja2 = "^3.1.2"
python-dotenv = "^0.21.0"
django-auth-adfs = "^1.11.1"
pip-system-certs = "^4.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
pygments = "^2.14.0"
django-filter = "^22.1"
django-guardian = "^2.4.0"
weasyprint = "^57.2"
python-barcode = "^0.14.0"
qrcode = "^7.3.1"
django-split-settings = "^1.2.0"
django-tinymce = "^3.5.0"
uwsgi = {version = "^2.0.21", platform = "linux"}
psycopg2-binary = "^2.9.5"
mssql-django = "^1.2"


[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.17.0"
mypy = "^0.991"
pre-commit = "^2.21.0"
ipython = "^8.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
